fix: update /info to not require auth
